refactor: Extract use hierarchy parents
